Move errdefs => api/errdefs
