+++ aci/context.go
@@ -28,7 +28,7 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/docker/compose-cli/api/context/store"
-	"github.com/docker/compose-cli/errdefs"
+	"github.com/docker/compose-cli/api/errdefs"
 	"github.com/docker/compose-cli/utils/prompt"
 )
