+++ cli/cmd/context/create_ecs.go
@@ -28,8 +28,8 @@ import (
 
 	"github.com/docker/compose-cli/api/client"
 	"github.com/docker/compose-cli/api/context/store"
+	"github.com/docker/compose-cli/api/errdefs"
 	"github.com/docker/compose-cli/ecs"
-	"github.com/docker/compose-cli/errdefs"
 )
 
 func init() {
