+++ ecs/local/context.go
@@ -20,8 +20,8 @@ import (
 	"context"
 
 	"github.com/docker/compose-cli/api/cloud"
+	"github.com/docker/compose-cli/api/errdefs"
 	"github.com/docker/compose-cli/ecs"
-	"github.com/docker/compose-cli/errdefs"
 )
 
 var _ cloud.Service = ecsLocalSimulation{}
