+++ local/containers.go
@@ -34,7 +34,7 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/docker/compose-cli/api/containers"
-	"github.com/docker/compose-cli/errdefs"
+	"github.com/docker/compose-cli/api/errdefs"
 	"github.com/docker/compose-cli/local/moby"
 )
